requirements.txt
Flask
mysql-connector-python
requests
gunicorn
flask
mysql-connector-python
requests
gunicorn
werkzeug
flask-wtf
flask-limiter
